requirements.txt
anthropic
pandas
python-dotenv
numpy
matplotlib
seaborn